feat(app/payroll-cycles/+page.server.ts): add support for showing closed payroll cycles
This commit adds support for showing closed payroll cycles on the payroll cycles page. The user can now choose to view all payroll cycles, including closed ones, by adding the `closed=true` query parameter to the URL. This allows for more flexibility in viewing payroll cycle data.

feat(app/payroll-cycles/+page.svelte): add showClosedCycles prop to page

This commit adds a new prop to the payroll cycles page called `showClosedCycles`. This prop is used to determine whether or not to show closed payroll cycles in the table. This allows for more flexibility in how the page is displayed.