feat(app): add reopen functionality to payroll cycles
Adds a button to reopen a closed payroll cycle. This allows users to make changes to paystubs after the cycle has been closed.

The reopen functionality is implemented using a modal that confirms the user's intent to reopen the cycle.

The reopen functionality is only available if the payment date for the cycle is in the future. This prevents users from reopening cycles that have already been paid out.